pyproject.toml
[tool.poetry]
name = "tcc-rastreamento-back"
version = "0.1.0"
description = "Backend do projeto de tcc em python"
authors = ["Henrique Franco Herrmann"]
readme = "README.md"
packages = [
  { include = "core" },
  { include = "utils" },
  { include = "middleware" }
]

[tool.poetry.dependencies]
python = "^3.13"
fastapi = ">=0.116.1,<0.117.0"
sqlalchemy = ">=2.0.43,<3.0.0"
asyncpg = ">=0.30.0,<0.31.0"
psycopg2-binary = ">=2.9.10,<3.0.0"
passlib = { version = ">=1.7.4,<2.0.0", extras = ["bcrypt"] }
python-jose = { version = ">=3.5.0,<4.0.0", extras = ["cryptography"] }
pydantic-settings = ">=2.10.1,<3.0.0"
alembic = ">=1.16.5,<2.0.0"
sqlalchemy-continuum = ">=1.5.0,<2.0.0"
pyyaml = ">=6.0.3,<7.0.0"
uvicorn = { version = "^0.30.0", extras = ["standard"] }
debugpy = "^1.8.17"
pytz = "^2025.2"

[tool.poetry.group.dev.dependencies]
pytest = "^8.0.0"

[build-system]
requires = ["poetry-core>=1.9.0"]
build-backend = "poetry.core.masonry.api"